wrap: no more implicit casts

+++ async/dns.C
@@ -722,7 +722,8 @@ dnsreq_ptr::readreply (dnsparse *reply)
       remove ();
       for (u_int i = 0; i < names.size (); i++)
 	vrfyv.push_back (New dnsreq_a (resp, names[i],
-				       wrap (this, &dnsreq_ptr::readvrfy, i),
+                                   wrap (this, &dnsreq_ptr::readvrfy,
+                                         static_cast<int>(i)),
 				       addr));
       return;
     }
